requirements.txt
playwright==1.29.1
marshmallow-dataclass==8.5.11
pytest-playwright==0.3.0
coverage==7.1.0